dentistry-project/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==4.2.14
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-yasg==1.21.7
flake8==7.1.1
gunicorn==23.0.0
idna==3.8
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==2.1.5
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pip-check-reqs==2.5.3
pluggy==1.5.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pytest==7.1.2
pytest-django==4.9.0
pytest-lazy-fixture==0.6.3
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
setuptools==75.2.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.4
sqlparse==0.5.1
tomli==2.0.2
transliterate==1.10.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
